Initial commit of pyproject-toml branch
This branch moves much of the configuration of the cyipopt package to
pyproject.toml. In particular, the dependencies are properly accounted
for so that cyipopt can be used in a poetry environment. It should also
reduce failures of `pip install cyipopt` that I have encountered. In 
addition, it greatly simplifies the setp.py script, which now only
has to generate the extension modules.