Use Cython <3 legacy behavior.
Cython 3 causes build errors due to the new requirement of the noexcept
clause. The language_level compiler directive is also set to Python 3 by
default. This commit restores the legacy behavior so the builds work
with Cython <3 and >=3. We will need to make proper fixes, as Cython
will probably drop the legacy functionality at some point.