Draft changes to improve exception handling
The changes here produce generally better error messages for cyipopt. In
many cases, errors in the callbacks produced predictable errors in
Ipopt, but with error messages that would be somewhat cryptic. So for
examples, errors in the hessianstructure callback produced status code
-12 errors (Invalid Option), with the invalid option being
hessian_approximation = “limited-memory”, which from a user perspective
does not point to an error in the hessianstructure callback. With these
changes, errors in the structure callbacks or exceptions in the other
callbacks all produce meaningful exceptions and error messages.